Fix data integrity issues
I feel like I have to be using Firebase wrong.
The previous state would blow up when trying to toggle a vote on
an item that had been voted on by someone who had not been loaded -
this would be because they had not actually submitted an idea or a
comment. This goes back to loading the votes based on the current user,
whose votes we're sure to have loaded, and ensures that when an idea
or comment are deleted, the associated records are also deleted.